% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gps_data.R
\name{add_gps_data}
\alias{add_gps_data}
\title{Joins the sf dataframe with the excel dataframe that has the salinity data}
\usage{
add_gps_data(x, y = salinity_master)
}
\arguments{
\item{x}{the sf dataframe}

\item{y}{the salinity data dataframe.  Defaults to salinity_master}
}
\value{
an sf dataframe
}
\description{
Joins the sf dataframe with the excel dataframe that has the salinity data
}
