% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gpx.R
\name{load_gpx}
\alias{load_gpx}
\title{Load the GPX files
pulls GPX files from the waypoints folder}
\usage{
load_gpx(x)
}
\arguments{
\item{x}{a file name that you want to load without the .gpx}
}
\value{
an sf dataframe
}
\description{
Load the GPX files
pulls GPX files from the waypoints folder
}
