% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shapefile.R
\name{make_shapefile}
\alias{make_shapefile}
\title{Write new sf dataframe as a shapefile
writes shapefile to the data folder}
\usage{
make_shapefile(r_name, shp_name)
}
\arguments{
\item{r_name}{the name of the sf dataframe you have been using in R}

\item{shp_name}{the name that you want the shapefile to be saved as}
}
\value{
a shapefile
}
\description{
Write new sf dataframe as a shapefile
writes shapefile to the data folder
}
