% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_factor.R
\name{no_factor}
\alias{no_factor}
\title{Converts the waypoint column from a factor to a number so you can join it later}
\usage{
no_factor(x)
}
\arguments{
\item{x}{the sf dataframe that you loaded earlier}
}
\value{
an sf dataframe
}
\description{
Converts the waypoint column from a factor to a number so you can join it later
}
