% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity_format.R
\name{salinity_format}
\alias{salinity_format}
\title{Formats the GPX and Excel data in a way that combines the data and makes it useable
joins the data}
\usage{
salinity_format(x, y = salinity_master)
}
\arguments{
\item{x}{sf dataframe}

\item{y}{excel dataframe.  defaults to salinity_master}
}
\value{
an sf dataframe
}
\description{
Formats the GPX and Excel data in a way that combines the data and makes it useable
joins the data
}
