% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinity_wrangle.R
\name{salinity_wrangle}
\alias{salinity_wrangle}
\title{Wraps three lines of code into 1?}
\usage{
salinity_wrangle(x)
}
\arguments{
\item{x}{the filename in the waypoints folder must be in quotes}
}
\value{
an shapefile
}
\description{
Wraps three lines of code into 1?
}
