% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_gpx.R
\name{simplify_gpx}
\alias{simplify_gpx}
\title{Takes the GPX files and removes non essential columns}
\usage{
simplify_gpx(x)
}
\arguments{
\item{x}{the sf dataframe that you loaded earlier}
}
\value{
an sf dataframe with fewer columns
}
\description{
Takes the GPX files and removes non essential columns
}
