% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salinitykayak.R
\name{simplify_gpx}
\alias{simplify_gpx}
\title{kayak data adjuster}
\usage{
simplify_gpx(x)
}
\arguments{
\item{x}{The name of the gpx dataframe you want to append to}

\item{y}{The name of the excel dataframe you want to use. default is salinity_master}
}
\description{
This function adds gpx files and excel files for the salinity kayak survey
}
\examples{
salinity_format(gpx_dataframe, excel_dataframe)
}
\keyword{salinity}
